+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/PromptDelegateTest.kt
@@ -4,6 +4,7 @@ import org.mozilla.geckoview.AllowOrDeny
 import org.mozilla.geckoview.GeckoResult
 import org.mozilla.geckoview.GeckoSession
 import org.mozilla.geckoview.GeckoSession.NavigationDelegate.LoadRequest
+import org.mozilla.geckoview.GeckoSession.PromptDelegate
 import org.mozilla.geckoview.test.rule.GeckoSessionTestRule.AssertCalled
 import org.mozilla.geckoview.test.util.Callbacks
 
@@ -25,11 +26,11 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.waitUntilCalled(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onPopupRequest(session: GeckoSession, targetUri: String?)
-                    : GeckoResult<AllowOrDeny>? {
+            override fun onPopupPrompt(session: GeckoSession, prompt: PromptDelegate.PopupPrompt)
+                    : GeckoResult<PromptDelegate.PromptResponse>? {
                 assertThat("Session should not be null", session, notNullValue())
-                assertThat("URL should not be null", targetUri, notNullValue())
-                assertThat("URL should match", targetUri, endsWith(HELLO_HTML_PATH))
+                assertThat("URL should not be null", prompt.targetUri, notNullValue())
+                assertThat("URL should match", prompt.targetUri, endsWith(HELLO_HTML_PATH))
                 return null
             }
         })
@@ -41,12 +42,12 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.delegateDuringNextWait(object : Callbacks.PromptDelegate, Callbacks.NavigationDelegate {
             @AssertCalled(count = 1)
-            override fun onPopupRequest(session: GeckoSession, targetUri: String?)
-                    : GeckoResult<AllowOrDeny>? {
+            override fun onPopupPrompt(session: GeckoSession, prompt: PromptDelegate.PopupPrompt)
+                    : GeckoResult<PromptDelegate.PromptResponse>? {
                 assertThat("Session should not be null", session, notNullValue())
-                assertThat("URL should not be null", targetUri, notNullValue())
-                assertThat("URL should match", targetUri, endsWith(HELLO_HTML_PATH))
-                return GeckoResult.fromValue(AllowOrDeny.ALLOW)
+                assertThat("URL should not be null", prompt.targetUri, notNullValue())
+                assertThat("URL should match", prompt.targetUri, endsWith(HELLO_HTML_PATH))
+                return GeckoResult.fromValue(prompt.confirm(AllowOrDeny.ALLOW))
             }
 
             @AssertCalled(count = 2)
@@ -69,12 +70,12 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.delegateDuringNextWait(object : Callbacks.PromptDelegate, Callbacks.NavigationDelegate {
             @AssertCalled(count = 1)
-            override fun onPopupRequest(session: GeckoSession, targetUri: String?)
-                    : GeckoResult<AllowOrDeny>? {
+            override fun onPopupPrompt(session: GeckoSession, prompt: PromptDelegate.PopupPrompt)
+                    : GeckoResult<PromptDelegate.PromptResponse>? {
                 assertThat("Session should not be null", session, notNullValue())
-                assertThat("URL should not be null", targetUri, notNullValue())
-                assertThat("URL should match", targetUri, endsWith(HELLO_HTML_PATH))
-                return GeckoResult.fromValue(AllowOrDeny.DENY)
+                assertThat("URL should not be null", prompt.targetUri, notNullValue())
+                assertThat("URL should match", prompt.targetUri, endsWith(HELLO_HTML_PATH))
+                return GeckoResult.fromValue(prompt.confirm(AllowOrDeny.DENY))
             }
 
             @AssertCalled(count = 1)
@@ -102,8 +103,9 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.waitUntilCalled(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onAlert(session: GeckoSession, title: String?, msg: String?, callback: GeckoSession.PromptDelegate.AlertCallback) {
-                assertThat("Message should match", "Alert!", equalTo(msg))
+            override fun onAlertPrompt(session: GeckoSession, prompt: PromptDelegate.AlertPrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                assertThat("Message should match", "Alert!", equalTo(prompt.message))
+                return GeckoResult.fromValue(prompt.dismiss())
             }
         })
     }
@@ -113,8 +115,9 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.waitUntilCalled(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onAuthPrompt(session: GeckoSession, title: String?, msg: String?, options: GeckoSession.PromptDelegate.AuthOptions, callback: GeckoSession.PromptDelegate.AuthCallback) {
+            override fun onAuthPrompt(session: GeckoSession, prompt: PromptDelegate.AuthPrompt): GeckoResult<PromptDelegate.PromptResponse>? {
                 //TODO: Figure out some better testing here.
+                return null
             }
         })
     }
@@ -123,9 +126,9 @@ class PromptDelegateTest : BaseSessionTest() {
     @Test fun buttonTest() {
         sessionRule.delegateDuringNextWait(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onButtonPrompt(session: GeckoSession, title: String?, msg: String?, btnMsg: Array<out String>?, callback: GeckoSession.PromptDelegate.ButtonCallback) {
-                assertThat("Message should match", "Confirm?", equalTo(msg))
-                callback.confirm(GeckoSession.PromptDelegate.BUTTON_TYPE_POSITIVE)
+            override fun onButtonPrompt(session: GeckoSession, prompt: PromptDelegate.ButtonPrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                assertThat("Message should match", "Confirm?", equalTo(prompt.message))
+                return GeckoResult.fromValue(prompt.confirm(PromptDelegate.ButtonPrompt.Type.POSITIVE))
             }
         })
 
@@ -135,9 +138,9 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.delegateDuringNextWait(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onButtonPrompt(session: GeckoSession, title: String?, msg: String?, btnMsg: Array<out String>?, callback: GeckoSession.PromptDelegate.ButtonCallback) {
-                assertThat("Message should match", "Confirm?", equalTo(msg))
-                callback.confirm(GeckoSession.PromptDelegate.BUTTON_TYPE_NEGATIVE)
+            override fun onButtonPrompt(session: GeckoSession, prompt: PromptDelegate.ButtonPrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                assertThat("Message should match", "Confirm?", equalTo(prompt.message))
+                return GeckoResult.fromValue(prompt.confirm(PromptDelegate.ButtonPrompt.Type.NEGATIVE))
             }
         })
 
@@ -152,10 +155,10 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.delegateUntilTestEnd(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onTextPrompt(session: GeckoSession, title: String?, msg: String?, value: String?, callback: GeckoSession.PromptDelegate.TextCallback) {
-                assertThat("Message should match", "Prompt:", equalTo(msg))
-                assertThat("Default should match", "default", equalTo(value))
-                callback.confirm("foo")
+            override fun onTextPrompt(session: GeckoSession, prompt: PromptDelegate.TextPrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                assertThat("Message should match", "Prompt:", equalTo(prompt.message))
+                assertThat("Default should match", "default", equalTo(prompt.defaultValue))
+                return GeckoResult.fromValue(prompt.confirm("foo"))
             }
         })
 
@@ -175,7 +178,8 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.waitUntilCalled(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onChoicePrompt(session: GeckoSession, title: String?, msg: String?, type: Int, choices: Array<out GeckoSession.PromptDelegate.Choice>, callback: GeckoSession.PromptDelegate.ChoiceCallback) {
+            override fun onChoicePrompt(session: GeckoSession, prompt: PromptDelegate.ChoicePrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                return GeckoResult.fromValue(prompt.dismiss())
             }
         })
     }
@@ -188,9 +192,9 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.delegateDuringNextWait(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onColorPrompt(session: GeckoSession, title: String?, value: String?, callback: GeckoSession.PromptDelegate.TextCallback) {
-                assertThat("Value should match", "#ffffff", equalTo(value))
-                callback.confirm("#123456")
+            override fun onColorPrompt(session: GeckoSession, prompt: PromptDelegate.ColorPrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                assertThat("Value should match", "#ffffff", equalTo(prompt.defaultValue))
+                return GeckoResult.fromValue(prompt.confirm("#123456"))
             }
         })
 
@@ -225,7 +229,8 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.waitUntilCalled(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onDateTimePrompt(session: GeckoSession, title: String?, type: Int, value: String?, min: String?, max: String?, callback: GeckoSession.PromptDelegate.TextCallback) {
+            override fun onDateTimePrompt(session: GeckoSession, prompt: PromptDelegate.DateTimePrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                return GeckoResult.fromValue(prompt.dismiss())
             }
         })
     }
@@ -240,12 +245,14 @@ class PromptDelegateTest : BaseSessionTest() {
 
         sessionRule.waitUntilCalled(object : Callbacks.PromptDelegate {
             @AssertCalled(count = 1)
-            override fun onFilePrompt(session: GeckoSession, title: String?, type: Int, mimeTypes: Array<out String>?, capture: Int, callback: GeckoSession.PromptDelegate.FileCallback) {
-                assertThat("Length of mimeTypes should match", 2, equalTo(mimeTypes!!.size))
-                assertThat("First accept attribute should match", "image/*", equalTo(mimeTypes[0]))
-                assertThat("Second accept attribute should match", ".pdf", equalTo(mimeTypes[1]))
-                assertThat("Capture attribute should match", GeckoSession.PromptDelegate.CAPTURE_TYPE_USER, equalTo(capture))
+            override fun onFilePrompt(session: GeckoSession, prompt: PromptDelegate.FilePrompt): GeckoResult<PromptDelegate.PromptResponse> {
+                assertThat("Length of mimeTypes should match", 2, equalTo(prompt.mimeTypes!!.size))
+                assertThat("First accept attribute should match", "image/*", equalTo(prompt.mimeTypes?.get(0)))
+                assertThat("Second accept attribute should match", ".pdf", equalTo(prompt.mimeTypes?.get(1)))
+                assertThat("Capture attribute should match", PromptDelegate.FilePrompt.Capture.USER, equalTo(prompt.capture))
+                return GeckoResult.fromValue(prompt.dismiss())
             }
         })
     }
 }
+